Refactor: Remove unused partition CSV and update platformio.ini
This commit deletes the `partitions_app8_spiffs8.csv` file as it is no longer needed. Additionally, the `platformio.ini` file has been updated to comment out the environment configuration for `esp32-s3-devkitm-1` and to specify the correct board for `esp32-s3-devkitc-1-n16r8v`. The changes ensure a cleaner configuration and better alignment with the current project structure.